feat(auth): implement useLayoutAuth hook for user authentication management in MainLayout
