+++ frontend/src/common/layouts/MainLayout.jsx
@@ -1,42 +1,15 @@
-import React, { useMemo, useState } from "react";
+import React from "react";
 import { Outlet, useNavigate } from "react-router-dom";
 
 import { Header,
          Footer,
          ScrollToTopButton
        } from "@orif-informatique/react-components-library";
-import useAuthStore from "../../features/auth/authStore";
+import useLayoutAuth from "../hooks/useLayoutAuth";
 
 const MainLayout = () => {
   const navigate = useNavigate();
-  const [resetKey, setResetKey] = useState(0);
-  const user = useAuthStore((s) => s.user);
-  const clearAuth = useAuthStore((s) => s.clearAuth);
-
-  const handleLogout = () => {
-    clearAuth();
-    localStorage.removeItem('loginType');
-    navigate('/');
-    setResetKey((k) => k + 1); // force remount for a cleaner visual reset
-  };
-
-  const storedUser = useMemo(() => {
-    try {
-      const raw = localStorage.getItem('auth-storage');
-      if (!raw) return null;
-      const parsed = JSON.parse(raw);
-      return parsed?.state?.user || null;
-    } catch (e) {
-      return null;
-    }
-  }, []);
-
-  const effectiveUser = user || storedUser;
-
-  const headerChild = useMemo(() => {
-    if (!effectiveUser) return null;
-    return null;
-  }, [effectiveUser?.firstName, effectiveUser?.lastName, effectiveUser?.login]);
+  const { effectiveUser, resetKey, handleLogout } = useLayoutAuth();
 
 
   return (
@@ -47,7 +20,6 @@ const MainLayout = () => {
         logoPath="/images/logo.svg"
         onLogin={() => navigate('/login')}
         onLogout={handleLogout}
-        childElement={headerChild}
         user={effectiveUser ? { name: effectiveUser?.firstName || "", role: effectiveUser?.mainRole || "user" } : null}
       />
       <main className="p-5 sm:p-10 bg-background">
